backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
openai
openpyxl
lxml
